move get_stop_list_dict
